requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
docx
annotated-types
anyio
bcrypt
click
colorama
DAWG-Python
docopt
ecdsa
fastapi
greenlet
h11
idna
lxml
passlib
pip
pyasn1
pydantic
pydantic_core
PyJWT
pymorphy2
pymorphy2-dicts-ru
python-docx
python-jose
router
rsa
setuptools
six
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
